requirements.txt
Flask==2.0.2
Flask-SQLAlchemy==3.0.0
python-dotenv==0.19.1
pymysql==1.0.2